untested things, but updating user delete and logout functions
also, no longer dependent on organizations, so correct plan accordingly

+++ deploy/confirm_account_deletion.sql
@@ -26,36 +26,24 @@ begin
 
   if v_user_secret is null then
     -- Success: they're already deleted
+    -- but wierd, because session should be invalid.  so fix that
+    perform app_public.logout();
     return true;
   end if;
 
   -- Check the token
-  if v_user_secret.delete_account_token = token then
-    -- Token passes
-
-    -- Check that they are not the owner of any organizations
-    if exists(
-      select 1
-      from app_public.organization_memberships
-      where user_id = app_public.current_user_id()
-      and is_owner is true
+  if(
+      -- token is still valid
+      v_user_secret.delete_account_token_generated > now() - v_token_max_duration
+    and
+      -- token matches
+      v_user_secret.delete_account_token = token
     ) then
-      raise exception 'You cannot delete your account until you are not the owner of any organizations.' using errcode = 'OWNER';
-    end if;
-
-    -- Reassign billing contact status back to the organization owner
-    update app_public.organization_memberships
-      set is_billing_contact = true
-      where is_owner = true
-      and organization_id in (
-        select organization_id
-        from app_public.organization_memberships my_memberships
-        where my_memberships.user_id = app_public.current_user_id()
-        and is_billing_contact is true
-      );
-
+    -- Token passes
     -- Delete their account :(
     delete from app_public.users where id = app_public.current_user_id();
+    -- also invalidate session
+    perform app_public.logout();
     return true;
   end if;
